forum included now
Note the html for the table and buttons did not transfer well from lab 4, I have no idea why
