+++ debug.log
@@ -1,0 +1,1 @@
+[1206/182848.260:ERROR:directory_reader_win.cc(43)] FindFirstFile: The system cannot find the path specified. (0x3)
