+++ package.json
@@ -13,9 +13,11 @@
     "react": "^17.0.1",
     "react-compound-interest": "^1.3.1",
     "react-dom": "^17.0.1",
+    "react-facebook": "^8.1.4",
     "react-google-login": "^5.1.25",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.0",
+    "react-twitter-embed": "^3.0.3",
     "web-vitals": "^0.2.4"
   },
   "author": "UF QuaranTeens",
