+++ src/App.css
@@ -1,3 +1,4 @@
+  
 .App {
   text-align: center;
 }
@@ -41,7 +42,7 @@
 /*NAVIGATION BAR*/
 
 .topnav {
-  background-color: #ff6600;
+  background-color: #dd6f08;
   overflow: hidden;
 }
 
@@ -53,24 +54,32 @@
   padding: 25px 25px;
   text-decoration: none;
   font-size: 44px;
+  font-weight: bold;
 }
 
 /* Change the color of links on hover */
 .topnav a:hover {
   background-color: #ddd;
   color: black;
+  text-decoration: underline  #dd6f08;
 }
 
 /* UF logo */
 .topnav a.active {
-  background-color: #0033ff;
+  background-color: #1069b6;
   color: white;
+}
+
+.topnav a.active:hover {
+  background-color: #0e5fa5;
+  color: white;
+  text-decoration: underline  #0e5fa5;
 }
 
 
 /* DISCLAIMER TEXT*/
 .page_disclaim {
-  background-color: #0033ff;
+  background-color: #1069b6;
   float: right;
   text-align: center;
   color: white;
@@ -83,7 +92,7 @@
   border-style: solid;
   padding: 1px;
   text-align: center;
-  background: #0033ff;
+  background: #1069b6;
   color: white;
   font-size: 15px;
 
@@ -102,6 +111,7 @@
   float: left;
   width: 50%;
   padding: 10px;
+
 }
 
 /* Clear floats after the columns */
