+++ src/Components/Academics.js
@@ -1,0 +1,49 @@
+import React from 'react';
+
+export default function Academics() {
+        return (
+            <div classname="AcademicsPage">
+              <div class="column">
+                <h1>Undergraduate Resources</h1>
+                <div class="column">
+                  <h2>Advising</h2>
+                  <p><a href="http://www.ufadvising.ufl.edu/college-remote-advising-contacts/"
+                  target="_blank" rel="noopener noreferrer">College Remote Advising Center</a></p>
+                </div>
+                <div class="column">
+                  <h2>Tutoring</h2>
+                  <p><a href="https://teachingcenter.ufl.edu/tutoring/" target="_blank"
+                  rel="noopener noreferrer">Teaching Center</a></p>
+                </div>
+                <div class="column">
+                  <h2>Job Resources</h2>
+                  <p><a href="https://career.ufl.edu/"
+                  target="_blank" rel="noopener noreferrer">Career Center</a></p>
+                </div>
+                <div class="column">
+                  <h2>Honors Programs</h2>
+                  <p><a href="http://www.honors.ufl.edu/"
+                  target="_blank" rel="noopener noreferrer">Honors Program</a></p>
+                </div>
+                <div class="column">
+                  <h2>Admissions</h2>
+                  <p><a href="https://admissions.ufl.edu/apply/transfer/"
+                  target="_blank" rel="noopener noreferrer">Transfer Student Center</a></p>
+                </div>
+                <div class="column">
+                  <h2>Scholarships</h2>
+                  <p><a href="https://cur.aa.ufl.edu/scholarships/"
+                  target="_blank" rel="noopener noreferrer">Center for Undergraduate Research</a></p>
+                </div>
+              </div>
+              <div class="column">
+                <h1>Graduate Resources</h1>
+                <div class="column">
+                  <h2>Graduate Resources</h2>
+                  <p><a href="http://graduateschool.ufl.edu/prospective-students/funding/fellowships/"
+                  target="_blank" rel="noopener noreferrer">Fellowships Center</a></p>
+                </div>
+              </div>
+            </div>
+          );
+}
