+++ src/Components/Map.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import GoogleMap, {addMarkers, locationList, mapOptions} from './GoogleMapAPI.js';
+
+export default function Map() {
+    var mapProps = {
+    options: mapOptions,
+    onMount: addMarkers,
+    onMountProps: locationList,
+  }
+  
+  return (
+    <div>
+      <div>
+      <h1>Map</h1>
+      </div>
+      <GoogleMap{...mapProps} />
+    </div>
+  );
+  }
